+++ dotnet/src/Connectors/Connectors.HuggingFace.UnitTests/HuggingFaceKernelBuilderExtensionsTests.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.Connectors.HuggingFace;
 using Microsoft.SemanticKernel.Embeddings;
@@ -26,19 +25,6 @@ public class HuggingFaceKernelBuilderExtensionsTests
     }
 
     [Fact]
-    public void AddHuggingFaceTextGenerationToServiceCollection()
-    {
-        var services = new ServiceCollection();
-        services.AddHuggingFaceTextGeneration("model");
-
-        var serviceProvider = services.BuildServiceProvider();
-        var service = serviceProvider.GetRequiredService<ITextGenerationService>();
-
-        Assert.NotNull(service);
-        Assert.IsType<HuggingFaceTextGenerationService>(service);
-    }
-
-    [Fact]
     public void AddHuggingFaceTextEmbeddingGenerationCreatesService()
     {
         var builder = Kernel.CreateBuilder();
@@ -48,19 +34,6 @@ public class HuggingFaceKernelBuilderExtensionsTests
         var service = kernel.GetRequiredService<ITextEmbeddingGenerationService>();
 
         Assert.NotNull(kernel);
-        Assert.NotNull(service);
-        Assert.IsType<HuggingFaceTextEmbeddingGenerationService>(service);
-    }
-
-    [Fact]
-    public void AddHuggingFaceTextEmbeddingsGenerationToServiceCollection()
-    {
-        var services = new ServiceCollection();
-        services.AddHuggingFaceTextEmbeddingGeneration("model");
-
-        var serviceProvider = services.BuildServiceProvider();
-        var service = serviceProvider.GetRequiredService<ITextEmbeddingGenerationService>();
-
         Assert.NotNull(service);
         Assert.IsType<HuggingFaceTextEmbeddingGenerationService>(service);
     }
