+++ dotnet/src/Connectors/Connectors.HuggingFace/Client/TextEmbeddingRequest.cs
@@ -3,7 +3,7 @@
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
-namespace Microsoft.SemanticKernel.Connectors.HuggingFace.Core;
+namespace Microsoft.SemanticKernel.Connectors.HuggingFace.Client;
 
 /// <summary>
 /// HTTP schema to perform embedding request.
