+++ dotnet/src/Connectors/Connectors.HuggingFace/Client/TextGenerationResponse.cs
@@ -2,11 +2,11 @@
 
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
-using static Microsoft.SemanticKernel.Connectors.HuggingFace.Core.TextGenerationResponse;
+using static Microsoft.SemanticKernel.Connectors.HuggingFace.Client.TextGenerationResponse;
 
 #pragma warning disable CA1812 // Avoid uninstantiated internal classes
 
-namespace Microsoft.SemanticKernel.Connectors.HuggingFace.Core;
+namespace Microsoft.SemanticKernel.Connectors.HuggingFace.Client;
 
 internal sealed class TextGenerationResponse : List<GeneratedTextItem>
 {
