+++ dotnet/src/Connectors/Connectors.HuggingFace/Client/TextGenerationStreamJsonParser.cs
@@ -5,9 +5,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Text.Json.Nodes;
-using Microsoft.SemanticKernel.Connectors.HuggingFace.Core;
 
-namespace Microsoft.SemanticKernel.Connectors.HuggingFace.TextGeneration;
+namespace Microsoft.SemanticKernel.Connectors.HuggingFace.Client;
 
 internal sealed class TextGenerationStreamJsonParser : IStreamJsonParser
 {
