--- dotnet/src/Connectors/Connectors.HuggingFace/TextGeneration/HuggingFaceTextGenerationService.cs
@@ -1,63 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-
-using System;
-using System.Collections.Generic;
-using System.Net.Http;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions;
-using Microsoft.SemanticKernel.Connectors.HuggingFace.Core;
-using Microsoft.SemanticKernel.Http;
-using Microsoft.SemanticKernel.Services;
-using Microsoft.SemanticKernel.TextGeneration;
-
-namespace Microsoft.SemanticKernel.Connectors.HuggingFace;
-
-/// <summary>
-/// HuggingFace text generation service.
-/// </summary>
-public sealed class HuggingFaceTextGenerationService : ITextGenerationService
-{
-    private Dictionary<string, object?> AttributesInternal { get; } = new();
-    private HuggingFaceClient Client { get; }
-
-    /// <inheritdoc />
-    public IReadOnlyDictionary<string, object?> Attributes => this.AttributesInternal;
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="HuggingFaceTextGenerationService"/> class.
-    /// </summary>
-    /// <param name="model">The HuggingFace model for the text generation service.</param>
-    /// <param name="endpoint">The uri endpoint including the port where HuggingFace server is hosted</param>
-    /// <param name="apiKey">Optional API key for accessing the HuggingFace service.</param>
-    /// <param name="httpClient">Optional HTTP client to be used for communication with the HuggingFace API.</param>
-    /// <param name="loggerFactory">Optional logger factory to be used for logging.</param>
-    public HuggingFaceTextGenerationService(
-        string model,
-        Uri? endpoint = null,
-        string? apiKey = null,
-        HttpClient? httpClient = null,
-        ILoggerFactory? loggerFactory = null)
-    {
-        Verify.NotNullOrWhiteSpace(model);
-
-        this.Client = new HuggingFaceClient(
-        modelId: model,
-            endpoint: endpoint ?? httpClient?.BaseAddress,
-            apiKey: apiKey,
-            httpClient: HttpClientProvider.GetHttpClient(httpClient),
-            logger: loggerFactory?.CreateLogger(this.GetType()) ?? NullLogger.Instance
-        );
-
-        this.AttributesInternal.Add(AIServiceExtensions.ModelIdKey, model);
-    }
-
-    /// <inheritdoc />
-    public Task<IReadOnlyList<TextContent>> GetTextContentsAsync(string prompt, PromptExecutionSettings? executionSettings = null, Kernel? kernel = null, CancellationToken cancellationToken = default)
-        => this.Client.GenerateTextAsync(prompt, executionSettings, cancellationToken);
-
-    /// <inheritdoc />
-    public IAsyncEnumerable<StreamingTextContent> GetStreamingTextContentsAsync(string prompt, PromptExecutionSettings? executionSettings = null, Kernel? kernel = null, CancellationToken cancellationToken = default)
-        => this.Client.StreamGenerateTextAsync(prompt, executionSettings, cancellationToken);
-}
